@@ -1,20 +1,16 @@
 package com.example
 
-import android.support.test.rule.ActivityTestRule
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Assert.fail
 import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
 @Ignore
 class ClassIgnoredTest {
 
-    @Rule
-    @JvmField
-    val rule = ActivityTestRule(MainActivity::class.java)
+    @get:Rule
+    val rule = ActivityScenarioRule(MainActivity::class.java)
 
     val screen = MainScreen()
 
