@@ -1,34 +1,28 @@
 package com.example
 
-import android.support.test.rule.ActivityTestRule
-import android.support.test.runner.AndroidJUnit4
-import io.qameta.allure.Epic
-import io.qameta.allure.Feature
-import io.qameta.allure.Owner
-import io.qameta.allure.Severity
-import io.qameta.allure.SeverityLevel
-import io.qameta.allure.Story
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import io.qameta.allure.android.runners.AllureAndroidJUnit4
+import io.qameta.allure.kotlin.*
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 
+@RunWith(AllureAndroidJUnit4::class)
 @Epic("General")
 @Feature("Text on main screen")
 @Story("Slow")
 @Owner("user2")
 @Severity(SeverityLevel.CRITICAL)
-@RunWith(AndroidJUnit4::class)
 class MainActivityTest {
 
-    @Rule
-    @JvmField
-    val rule = ActivityTestRule(MainActivity::class.java)
+    @get:Rule
+    val rule = ActivityScenarioRule(MainActivity::class.java)
 
     val screen = MainScreen()
 
     @Test
     fun testText() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -38,7 +32,7 @@ class MainActivityTest {
 
     @Test
     fun testText1() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -48,7 +42,7 @@ class MainActivityTest {
 
     @Test
     fun testText2() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -58,7 +52,7 @@ class MainActivityTest {
 
     @Test
     fun testText3() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -68,7 +62,7 @@ class MainActivityTest {
 
     @Test
     fun testText4() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -78,7 +72,7 @@ class MainActivityTest {
 
     @Test
     fun testText5() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -88,7 +82,7 @@ class MainActivityTest {
 
     @Test
     fun testText6() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -98,7 +92,7 @@ class MainActivityTest {
 
     @Test
     fun testText7() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -108,7 +102,7 @@ class MainActivityTest {
 
     @Test
     fun testText8() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
@@ -118,7 +112,7 @@ class MainActivityTest {
 
     @Test
     fun testText9() {
-        java.lang.Thread.sleep(500)
+        Thread.sleep(500)
         screen {
             text {
                 hasText("Test")
