@@ -1,17 +1,18 @@
 object Versions {
-    val kotlin = "1.3.61"
-    val coroutines = "1.3.3"
+    val kotlin = "1.4.10"
+    val coroutines = "1.3.9"
 
-    val androidGradleVersion = "3.6.2"
+    val androidGradleVersion = "4.0.0"
 
-    val kakao = "1.2.1"
-    val espresso = "3.0.1"
-    val espressoRules = "1.0.1"
-    val espressoRunner = "1.0.1"
+    val kakao = "2.4.0"
+    val espresso = "3.3.0"
+    val espressoRules = "1.3.0"
+    val espressoRunner = "1.3.0"
+    val testJunit = "1.1.2"
     val junit = "4.12"
-    val appCompat = "28.0.0"
-    val constraintLayout = "1.0.2"
-    val allure = "2.8.1"
+    val appCompat = "1.2.0"
+    val constraintLayout = "2.0.4"
+    val allure = "2.1.2"
 
 }
 
@@ -22,21 +23,20 @@ object BuildPlugins {
 
 object Libraries {
     val kotlinStdLib = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.kotlin}"
-    val kotlinReflect = "org.jetbrains.kotlin:kotlin-reflect:${Versions.kotlin}"
     val kotlinCoroutines = "org.jetbrains.kotlinx:kotlinx-coroutines-core:${Versions.coroutines}"
-    val appCompat = "com.android.support:appcompat-v7:${Versions.appCompat}"
-    val constraintLayout = "com.android.support.constraint:constraint-layout:${Versions.constraintLayout}"
+    val appCompat = "androidx.appcompat:appcompat:${Versions.appCompat}"
+    val constraintLayout = "androidx.constraintlayout:constraintlayout:${Versions.constraintLayout}"
 }
 
 object TestLibraries {
     val kakao = "com.agoda.kakao:kakao:${Versions.kakao}"
 
-    val espressoRunner = "com.android.support.test:runner:${Versions.espressoRunner}"
-    val espressoRules = "com.android.support.test:rules:${Versions.espressoRules}"
-    val espressoCore = "com.android.support.test.espresso:espresso-core:${Versions.espresso}"
-    val espressoWeb = "com.android.support.test.espresso:espresso-web:${Versions.espresso}"
-    val espressoContrib = "com.android.support.test.espresso:espresso-contrib:${Versions.espresso}"
-    val espressoIntents = "com.android.support.test.espresso:espresso-intents:${Versions.espresso}"
-    val junit = "junit:junit:${Versions.junit}"
-    val allure = "io.qameta.allure:allure-java-commons:${Versions.allure}"
+    val testRunner = "androidx.test:runner:${Versions.espressoRunner}"
+    val testRules = "androidx.test:rules:${Versions.espressoRules}"
+    val extJunit = "androidx.test.ext:junit:${Versions.testJunit}"
+    val espressoCore = "androidx.test.espresso:espresso-core:${Versions.espresso}"
+    val allureKotlinModel = "io.qameta.allure:allure-kotlin-model:${Versions.allure}"
+    val allureKotlinCommon = "io.qameta.allure:allure-kotlin-commons:${Versions.allure}"
+    val allureKotlinJunit4 = "io.qameta.allure:allure-kotlin-junit4:${Versions.allure}"
+    val allureKotlinAndroid = "io.qameta.allure:allure-kotlin-android:${Versions.allure}"
 }
