@@ -1,22 +1,23 @@
 object Versions {
-    val marathon = System.getenv("DEPLOY_VERSION_OVERRIDE") ?: "0.5.2"
+    val marathon = System.getenv("DEPLOY_VERSION_OVERRIDE") ?: "0.7.0"
 
-    val kotlin = "1.3.61"
-    val coroutines = "1.3.1"
+    val kotlin = "1.4.20"
+    val coroutines = "1.3.9"
 
-    val ddmlib = "26.3.0"
+    val ddmlib = "27.1.2"
+    val adam = "0.2.3"
     val dexTestParser = "2.1.1"
-    val kotlinLogging = "1.4.9"
+    val kotlinLogging = "1.7.6"
     val slf4jAPI = "1.0.0"
     val logbackClassic = "1.2.3"
     val axmlParser = "1.0"
     val bugsnag = "3.6.1"
 
     val junitGradle = "1.0.0"
-    val androidGradleVersion = "3.6.1"
+    val androidGradleVersion = "4.0.0"
 
     val junit5 = "5.6.0"
-    val kluent = "1.40"
+    val kluent = "1.64"
 
     val kakao = "1.4.0"
     val espresso = "3.0.1"
@@ -36,18 +37,19 @@ object Versions {
     val guava = "26.0-jre"
     val rsync4j = "3.1.2-12"
     val sshj = "0.26.0"
-    val testContainers = "1.9.1"
+    val testContainers = "1.15.0-rc2"
     val jupiterEngine = "5.1.0"
     val jansi = "1.17.1"
     val scalr = "4.2"
-    val allure = "2.8.1"
+    val allureJava = "2.8.1"
+    val allureKotlin = "2.1.2"
     val allureEnvironment = "1.0.0"
     val mockitoKotlin = "2.0.0"
     val googleAnalitycsWrapper = "2.0.0"
     val dokka = "0.9.17"
     val koin = "2.0.1"
     val jsonAssert = "1.5.0"
-    val xmlUnit ="2.6.3"
+    val xmlUnit = "2.6.3"
     val assertk = "0.19"
 }
 
@@ -60,6 +62,7 @@ object BuildPlugins {
 
 object Libraries {
     val ddmlib = "com.android.tools.ddms:ddmlib:${Versions.ddmlib}"
+    val adam = "com.malinskiy:adam:${Versions.adam}"
     val androidCommon = "com.android.tools:common:${Versions.ddmlib}"
     val dexTestParser = "com.linkedin.dextestparser:parser:${Versions.dexTestParser}"
     val kotlinStdLib = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.kotlin}"
@@ -86,8 +89,9 @@ object Libraries {
     val sshj = "com.hierynomus:sshj:${Versions.sshj}"
     val jansi = "org.fusesource.jansi:jansi:${Versions.jansi}"
     val scalr = "org.imgscalr:imgscalr-lib:${Versions.scalr}"
-    val allure = "io.qameta.allure:allure-java-commons:${Versions.allure}"
+    val allure = "io.qameta.allure:allure-java-commons:${Versions.allureJava}"
     val allureEnvironment = "com.github.automatedowl:allure-environment-writer:${Versions.allureEnvironment}"
+    val allureKotlinCommons = "io.qameta.allure:allure-kotlin-commons:${Versions.allureKotlin}"
     val koin = "org.koin:koin-core:${Versions.koin}"
     val bugsnag = "com.bugsnag:bugsnag:${Versions.bugsnag}"
 }
@@ -97,12 +101,6 @@ object TestLibraries {
     val kluent = "org.amshove.kluent:kluent:${Versions.kluent}"
     val kakao = "com.agoda.kakao:kakao:${Versions.kakao}"
 
-    val espressoRunner = "com.android.support.test:runner:${Versions.espressoRunner}"
-    val espressoRules = "com.android.support.test:rules:${Versions.espressoRules}"
-    val espressoCore = "com.android.support.test.espresso:espresso-core:${Versions.espresso}"
-    val espressoWeb = "com.android.support.test.espresso:espresso-web:${Versions.espresso}"
-    val espressoContrib = "com.android.support.test.espresso:espresso-contrib:${Versions.espresso}"
-    val espressoIntents = "com.android.support.test.espresso:espresso-intents:${Versions.espresso}"
     val junit = "junit:junit:${Versions.junit}"
     val mockitoKotlin = "com.nhaarman.mockitokotlin2:mockito-kotlin:${Versions.mockitoKotlin}"
     val jupiterEngine = "org.junit.jupiter:junit-jupiter-engine:${Versions.jupiterEngine}"
