@@ -4,6 +4,7 @@ import com.malinskiy.marathon.device.DeviceInfo
 import com.malinskiy.marathon.device.DevicePoolId
 import com.malinskiy.marathon.execution.TestResult
 import com.malinskiy.marathon.execution.TestStatus
+import com.malinskiy.marathon.test.toTestName
 import java.time.Instant
 
 /**
@@ -31,28 +32,46 @@ data class ExecutionReport(
             .map { it.testResult }
             .filter { it.status != TestStatus.INCOMPLETE }
 
-        val passed = tests.count { it.status == TestStatus.PASSED }
-        val ignored = tests.count {
-            it.status == TestStatus.IGNORED
+        val passed = tests
+            .filter { it.status == TestStatus.PASSED }
+            .map { it.test.toTestName() }
+            .toSet()
+
+        val ignored = tests
+            .filter { it.status == TestStatus.IGNORED
                     || it.status == TestStatus.ASSUMPTION_FAILURE
-        }
-        val failed = tests.count {
-            it.status != TestStatus.PASSED
+            }.map { it.test.toTestName() }
+            .toSet()
+
+        val failed = tests
+            .filter { it.status != TestStatus.PASSED
                     && it.status != TestStatus.IGNORED
                     && it.status != TestStatus.ASSUMPTION_FAILURE
-        }
+             }.map { it.test.toTestName() }
+            .toSet()
+
         val duration = tests.map { it.durationMillis() }.sum()
 
         val rawTests = poolTestEvents
             .map { it.testResult }
-        val rawPassed = rawTests.count { it.status == TestStatus.PASSED }
-        val rawIgnored = rawTests.count {
-            it.status == TestStatus.IGNORED
-                    || it.status == TestStatus.ASSUMPTION_FAILURE
-        }
-        val rawFailed = rawTests.count { it.status == TestStatus.FAILURE }
-        val rawIncomplete = rawTests.count { it.status == TestStatus.INCOMPLETE }
-        val rawDuration = rawTests.map { it.durationMillis() }.sum()
+        val rawPassed = rawTests
+            .filter { it.status == TestStatus.PASSED }
+            .map { it.test.toTestName() }
+        val rawIgnored = rawTests
+            .filter { it.status == TestStatus.IGNORED
+                || it.status == TestStatus.ASSUMPTION_FAILURE
+            }.map { it.test.toTestName() }
+        val rawFailed = rawTests
+            .filter { it.status == TestStatus.FAILURE }
+            .map { it.test.toTestName() }
+        val rawIncomplete = rawTests
+            .filter { it.status == TestStatus.INCOMPLETE }
+            .map { it.test.toTestName() }
+        val rawDuration = rawTests
+            //Incomplete tests mess up the calculations of time since their end time is 0 and duration is, hence, years
+            //We filter here for unavailable time just to be safe
+            .filter { it.startTime != 0L && it.endTime != 0L }
+            .map { it.durationMillis() }.sum()
 
         return PoolSummary(
             poolId = poolId,
