@@ -7,15 +7,15 @@ import com.malinskiy.marathon.execution.TestResult
 data class PoolSummary(
     val poolId: DevicePoolId,
     val tests: List<TestResult>,
-    val passed: Int,
-    val ignored: Int,
-    val failed: Int,
+    val passed: Set<String>,
+    val ignored: Set<String>,
+    val failed: Set<String>,
     val flaky: Int,
     val durationMillis: Long,
     val devices: List<DeviceInfo>,
-    val rawPassed: Int,
-    val rawIgnored: Int,
-    val rawFailed: Int,
-    val rawIncomplete: Int,
+    val rawPassed: List<String>,
+    val rawIgnored: List<String>,
+    val rawFailed: List<String>,
+    val rawIncomplete: List<String>,
     val rawDurationMillis: Long
 )
