@@ -16,8 +16,9 @@ import com.malinskiy.marathon.execution.strategy.impl.sorting.NoSortingStrategy
 import com.malinskiy.marathon.vendor.VendorConfiguration
 import java.io.File
 
-private const val DEFAULT_EXECUTION_TIMEOUT_MILLIS: Long = 900_000
-private const val DEFAULT_OUTPUT_TIMEOUT_MILLIS: Long = 60_000
+private const val DEFAULT_BATCH_EXECUTION_TIMEOUT_MILLIS: Long = 1800_000 //30 min
+private const val DEFAULT_OUTPUT_TIMEOUT_MILLIS: Long = 300_000 //5 min
+private const val DEFAULT_DEVICE_INITIALIZATION_TIMEOUT_MILLIS = 180_000L
 
 data class Configuration constructor(
     val name: String,
@@ -50,7 +51,8 @@ data class Configuration constructor(
 
     val vendorConfiguration: VendorConfiguration,
 
-    val analyticsTracking: Boolean
+    val analyticsTracking: Boolean,
+    val deviceInitializationTimeoutMillis: Long
 ) {
 
     constructor(
@@ -84,7 +86,8 @@ data class Configuration constructor(
 
         vendorConfiguration: VendorConfiguration,
 
-        analyticsTracking: Boolean?
+        analyticsTracking: Boolean?,
+        deviceInitializationTimeoutMillis: Long?
     ) :
 
             this(
@@ -106,12 +109,13 @@ data class Configuration constructor(
                 testClassRegexes = testClassRegexes ?: listOf(Regex("^((?!Abstract).)*Test[s]*$")),
                 includeSerialRegexes = includeSerialRegexes ?: emptyList(),
                 excludeSerialRegexes = excludeSerialRegexes ?: emptyList(),
-                testBatchTimeoutMillis = testBatchTimeoutMillis ?: DEFAULT_EXECUTION_TIMEOUT_MILLIS,
+                testBatchTimeoutMillis = testBatchTimeoutMillis ?: DEFAULT_BATCH_EXECUTION_TIMEOUT_MILLIS,
                 testOutputTimeoutMillis = testOutputTimeoutMillis ?: DEFAULT_OUTPUT_TIMEOUT_MILLIS,
                 debug = debug ?: true,
                 screenRecordingPolicy = screenRecordingPolicy ?: ScreenRecordingPolicy.ON_FAILURE,
                 vendorConfiguration = vendorConfiguration,
-                analyticsTracking = analyticsTracking ?: false
+                analyticsTracking = analyticsTracking ?: false,
+                deviceInitializationTimeoutMillis = deviceInitializationTimeoutMillis ?: DEFAULT_DEVICE_INITIALIZATION_TIMEOUT_MILLIS
             )
 
     fun toMap() =
@@ -137,6 +141,7 @@ data class Configuration constructor(
             "testOutputTimeoutMillis" to testOutputTimeoutMillis.toString(),
             "debug" to debug.toString(),
             "screenRecordingPolicy" to screenRecordingPolicy.toString(),
-            "vendorConfiguration" to vendorConfiguration.toString()
+            "vendorConfiguration" to vendorConfiguration.toString(),
+            "deviceInitializationTimeoutMillis" to deviceInitializationTimeoutMillis.toString()
         )
 }
