@@ -15,7 +15,9 @@ import com.malinskiy.marathon.execution.queue.QueueActor
 import com.malinskiy.marathon.execution.queue.QueueMessage
 import com.malinskiy.marathon.log.MarathonLogging
 import com.malinskiy.marathon.test.TestBatch
+import com.malinskiy.marathon.time.Timer
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.channels.SendChannel
 import kotlin.coroutines.CoroutineContext
 
@@ -26,6 +28,7 @@ class DevicePoolActor(
     shard: TestShard,
     private val progressReporter: ProgressReporter,
     track: Track,
+    timer: Timer,
     parent: Job,
     context: CoroutineContext
 ) :
@@ -40,14 +43,17 @@ class DevicePoolActor(
             is DevicePoolMessage.FromScheduler.Terminate -> terminate()
             is DevicePoolMessage.FromDevice.IsReady -> deviceReady(msg)
             is DevicePoolMessage.FromDevice.CompletedTestBatch -> deviceCompleted(msg.device, msg.results)
-            is DevicePoolMessage.FromDevice.ReturnTestBatch -> deviceReturnedTestBatch(msg.device, msg.batch)
+            is DevicePoolMessage.FromDevice.ReturnTestBatch -> deviceReturnedTestBatch(msg.device, msg.batch, msg.reason)
             is DevicePoolMessage.FromQueue.Notify -> notifyDevices()
             is DevicePoolMessage.FromQueue.Terminated -> onQueueTerminated()
             is DevicePoolMessage.FromQueue.ExecuteBatch -> executeBatch(msg.device, msg.batch)
         }
     }
 
-    private val poolJob = Job(parent)
+    /**
+     * Any problem with a device should not propagate a cancellation upstream
+     */
+    private val poolJob = SupervisorJob(parent)
 
     private val queue: QueueActor = QueueActor(
         configuration,
@@ -57,6 +63,7 @@ class DevicePoolActor(
         poolId,
         progressReporter,
         track,
+        timer,
         poolJob,
         context
     )
@@ -77,8 +84,8 @@ class DevicePoolActor(
         terminate()
     }
 
-    private suspend fun deviceReturnedTestBatch(device: Device, batch: TestBatch) {
-        queue.send(QueueMessage.ReturnBatch(device.toDeviceInfo(), batch))
+    private suspend fun deviceReturnedTestBatch(device: Device, batch: TestBatch, reason: String) {
+        queue.send(QueueMessage.ReturnBatch(device.toDeviceInfo(), batch, reason))
     }
 
     private suspend fun deviceCompleted(device: Device, results: TestBatchResults) {
