@@ -27,6 +27,7 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.shouldBe
+import org.amshove.kluent.shouldBeInstanceOf
 import org.amshove.kluent.shouldContainSame
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
@@ -94,8 +95,9 @@ class QueueActorTest {
         val isEmptyDeferred = CompletableDeferred<Boolean>()
         runBlocking {
             actor.send(QueueMessage.RequestBatch(TEST_DEVICE_INFO))
-            poolChannel.receive()
+            poolChannel.receive() shouldBeInstanceOf FromQueue.ExecuteBatch::class
             actor.send(QueueMessage.Completed(TEST_DEVICE_INFO, testBatchResults))
+            poolChannel.receive() shouldBeInstanceOf FromQueue.Notify::class
             actor.send(QueueMessage.IsEmpty(isEmptyDeferred))
             isEmptyDeferred.await() shouldBe false
         }
@@ -107,8 +109,9 @@ class QueueActorTest {
 
         runBlocking {
             actor.send(QueueMessage.RequestBatch(TEST_DEVICE_INFO))
-            poolChannel.receive()
+            poolChannel.receive() shouldBeInstanceOf FromQueue.ExecuteBatch::class
             actor.send(QueueMessage.Completed(TEST_DEVICE_INFO, testBatchResults))
+            poolChannel.receive() shouldBeInstanceOf FromQueue.Notify::class
 
             verify(track, times(1)).test(any(), any(), testResultCaptor.capture(), any())
             testResultCaptor.firstValue.test shouldBe TEST_1
@@ -121,12 +124,14 @@ class QueueActorTest {
         setup_2()
         runBlocking {
             actor.send(QueueMessage.RequestBatch(TEST_DEVICE_INFO))
-            poolChannel.receive()
+            poolChannel.receive() shouldBeInstanceOf FromQueue.ExecuteBatch::class
             actor.send(QueueMessage.Completed(TEST_DEVICE_INFO, testBatchResults))
+            poolChannel.receive() shouldBeInstanceOf FromQueue.Notify::class
             actor.send(QueueMessage.RequestBatch(TEST_DEVICE_INFO))
-            val response = poolChannel.receive()
-            response::class shouldBe FromQueue.ExecuteBatch::class
-            (response as FromQueue.ExecuteBatch).batch.tests shouldContainSame listOf(TEST_1)
+            val actual = poolChannel.receive()
+
+            actual shouldBeInstanceOf FromQueue.ExecuteBatch::class
+            (actual as FromQueue.ExecuteBatch).batch.tests shouldContainSame listOf(TEST_1)
         }
     }
 
@@ -136,11 +141,12 @@ class QueueActorTest {
         val isEmptyDeferred = CompletableDeferred<Boolean>()
         runBlocking {
             actor.send(QueueMessage.RequestBatch(TEST_DEVICE_INFO))
-            poolChannel.receive()
+            poolChannel.receive() shouldBeInstanceOf FromQueue.ExecuteBatch::class
             actor.send(QueueMessage.Completed(TEST_DEVICE_INFO, testBatchResults))
 
             actor.send(QueueMessage.RequestBatch(TEST_DEVICE_INFO))
-            poolChannel.receive()
+            poolChannel.receive() shouldBeInstanceOf FromQueue.Notify::class
+            poolChannel.receive() shouldBeInstanceOf FromQueue.ExecuteBatch::class
             actor.send(QueueMessage.Completed(TEST_DEVICE_INFO, testBatchResults))
 
             actor.send(QueueMessage.IsEmpty(isEmptyDeferred))
@@ -253,6 +259,7 @@ private fun createQueueActor(
     DevicePoolId("test"),
     mock(),
     track,
+    mock(),
     job,
     Dispatchers.Unconfined
 )
@@ -284,5 +291,6 @@ private val DEFAULT_CONFIGURATION = Configuration(
         testParser = mock(),
         deviceProvider = mock()
     ),
-    analyticsTracking = false
+    analyticsTracking = false,
+    deviceInitializationTimeoutMillis = null
 )
