@@ -51,9 +51,9 @@ class ExecutionReportTest {
             override fun testParser(): TestParser? = null
             override fun deviceProvider(): DeviceProvider? = null
             override fun logConfigurator(): MarathonLogConfigurator? = null
-            override fun preferableRecorderType(): DeviceFeature? = null
         },
-        analyticsTracking = false
+        analyticsTracking = false,
+        deviceInitializationTimeoutMillis = null
     )
 
     private val reportWithoutRetries: ExecutionReport by lazy {
