@@ -210,9 +210,9 @@ fun getConfiguration() =
             override fun testParser(): TestParser? = null
             override fun deviceProvider(): DeviceProvider? = null
             override fun logConfigurator(): MarathonLogConfigurator? = null
-            override fun preferableRecorderType(): DeviceFeature? = null
         },
-        analyticsTracking = false
+        analyticsTracking = false,
+        deviceInitializationTimeoutMillis = null
     )
 
 fun getDevice() =
