@@ -1,11 +1,15 @@
 package com.malinskiy.marathon
 
 import com.malinskiy.marathon.execution.AnalyticsConfiguration
+import com.malinskiy.marathon.execution.AnalyticsConfiguration.DisabledAnalytics
+import com.malinskiy.marathon.execution.AnalyticsConfiguration.GraphiteConfiguration
+import com.malinskiy.marathon.execution.AnalyticsConfiguration.InfluxDbConfiguration
 import com.malinskiy.marathon.execution.AnalyticsConfiguration.InfluxDbConfiguration.RetentionPolicyConfiguration
 import groovy.lang.Closure
 
 class AnalyticsConfig {
     var influx: InfluxConfig? = null
+    var graphite: GraphiteConfig? = null
 
     fun influx(closure: Closure<*>) {
         influx = InfluxConfig()
@@ -15,6 +19,16 @@ class AnalyticsConfig {
 
     fun influx(block: InfluxConfig.() -> Unit) {
         influx = InfluxConfig().also(block)
+    }
+
+    fun graphite(closure: Closure<*>) {
+        graphite = GraphiteConfig()
+        closure.delegate = graphite
+        closure.call()
+    }
+
+    fun graphite(block: GraphiteConfig.() -> Unit) {
+        graphite = GraphiteConfig().also(block)
     }
 }
 
@@ -34,17 +48,31 @@ class RetentionPolicy {
     var isDefault: Boolean = true
 }
 
+class GraphiteConfig {
+    var host: String = ""
+    var port: String? = null
+    var prefix: String? = null
+}
+
 fun AnalyticsConfig.toAnalyticsConfiguration(): AnalyticsConfiguration {
-    return influx?.let {
-        AnalyticsConfiguration.InfluxDbConfiguration(
-            dbName = it.dbName,
-            user = it.user,
-            password = it.password,
-            url = it.url,
-            retentionPolicyConfiguration = it.retentionPolicy?.toRetentionPolicy()
+    val influx = this.influx
+    val graphite = this.graphite
+    return when {
+        influx != null -> InfluxDbConfiguration(
+            dbName = influx.dbName,
+            user = influx.user,
+            password = influx.password,
+            url = influx.url,
+            retentionPolicyConfiguration = influx.retentionPolicy?.toRetentionPolicy()
                 ?: RetentionPolicyConfiguration.default
         )
-    } ?: AnalyticsConfiguration.DisabledAnalytics
+        graphite != null -> GraphiteConfiguration(
+            host = graphite.host,
+            port = graphite.port?.toIntOrNull(),
+            prefix = graphite.prefix
+        )
+        else -> DisabledAnalytics
+    }
 }
 
 private fun RetentionPolicy.toRetentionPolicy(): RetentionPolicyConfiguration {
