@@ -1,13 +1,19 @@
 package com.malinskiy.marathon
 
-import com.malinskiy.marathon.android.serial.SerialStrategy
-import com.malinskiy.marathon.device.DeviceFeature
+import com.malinskiy.marathon.android.ScreenRecordConfiguration
+import com.malinskiy.marathon.android.VendorType
+import com.malinskiy.marathon.android.configuration.AllureConfiguration
+import com.malinskiy.marathon.android.configuration.SerialStrategy
+import com.malinskiy.marathon.android.configuration.TimeoutConfiguration
 import com.malinskiy.marathon.execution.policy.ScreenRecordingPolicy
 import groovy.lang.Closure
 import org.gradle.api.Project
 
 open class MarathonExtension(project: Project) {
     var name: String = "Marathon"
+
+    var vendor: VendorType? = null
+    var bugsnag: Boolean? = null
 
     var analyticsConfiguration: AnalyticsConfig? = null
 
@@ -35,7 +41,7 @@ open class MarathonExtension(project: Project) {
     var testOutputTimeoutMillis: Long? = null
     var debug: Boolean? = null
 
-    val screenRecordingPolicy: ScreenRecordingPolicy? = null
+    var screenRecordingPolicy: ScreenRecordingPolicy? = null
 
     var applicationPmClear: Boolean? = null
     var testApplicationPmClear: Boolean? = null
@@ -44,9 +50,15 @@ open class MarathonExtension(project: Project) {
     var serialStrategy: SerialStrategy? = null
     var serialStrategyName: String? = null
 
-    var preferableRecorderType: DeviceFeature? = null
+    var screenRecordConfiguration: ScreenRecordConfiguration? = null
 
     var analyticsTracking: Boolean = false
+
+    var deviceInitializationTimeoutMillis: Long? = null
+    var waitForDevicesTimeoutMillis: Long? = null
+
+    var allureConfiguration: AllureConfiguration? = null
+    var timeoutConfiguration: TimeoutConfiguration? = null
 
     //Android specific for now
     var autoGrantPermission: Boolean? = null
@@ -87,6 +99,14 @@ open class MarathonExtension(project: Project) {
 
     fun instrumentationArgs(block: MutableMap<String, String>.() -> Unit) {
         instrumentationArgs = mutableMapOf<String, String>().also(block)
+    }
+
+    fun allureConfiguration(block: AllureConfiguration.() -> Unit) {
+        allureConfiguration = AllureConfiguration().also(block)
+    }
+
+    fun timeoutConfiguration(block: TimeoutConfiguration.() -> Unit) {
+        timeoutConfiguration = TimeoutConfiguration().also(block)
     }
 
     //Groovy way
@@ -141,6 +161,18 @@ open class MarathonExtension(project: Project) {
     fun instrumentationArgs(closure: Closure<*>) {
         instrumentationArgs = mutableMapOf()
         closure.delegate = instrumentationArgs
+        closure.call()
+    }
+
+    fun allureConfiguration(closure: Closure<*>) {
+        allureConfiguration = AllureConfiguration()
+        closure.delegate = allureConfiguration
+        closure.call()
+    }
+
+    fun timeoutConfiguration(closure: Closure<*>) {
+        timeoutConfiguration = TimeoutConfiguration()
+        closure.delegate = timeoutConfiguration
         closure.call()
     }
 }
