@@ -1,7 +1,8 @@
 package com.example
 
-import com.agoda.kakao.KEditText
-import com.agoda.kakao.Screen
+import com.agoda.kakao.edit.KEditText
+import com.agoda.kakao.screen.Screen
+
 
 class MainScreen : Screen<MainScreen>() {
     val text = KEditText { withId(R.id.edittext) }
